Isolated classes Mongo, History and Rule
