+++ orakwlum/consumption/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 __author__ = 'XaviTorello'
 
-from consumption import *
+from consumption import Consumption
+from history import History
