+++ orakwlum/datasource/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 __author__ = 'XaviTorello'
 
-from datasource import *
+from datasource import DataSource
+from mongo import Mongo
