+++ orakwlum/scenario/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'XaviTorello'
 
-from scenario import *
+from scenario import Scenario, Rule
