+++ orakwlum/scenario/scenario.py
@@ -6,6 +6,7 @@ import logging
 logger = logging.getLogger(__name__)
 
 from orakwlum.consumption import History
+from orakwlum.scenario.rule import Rule
 
 
 class Scenario(object):
@@ -83,25 +84,3 @@ class Scenario(object):
         self.history.consumptions_hourly = self.history.get_consumption_hourly(
         )
         #self.history.dump_history_hourly()
-
-
-class Rule(object):
-    """
-    Create a new Rule related to a Scenario
-    """
-
-    def __init__(self, name, filter, filter_values, action, action_field,
-                 action_value):
-
-        assert filter and filter_values, "Filter is not defined"
-        assert action and action_field and action_value, "Action is not well defined"
-
-        if type(filter_values) == str:
-            filter_values = [filter_values]
-
-        self.name = name
-        self.filter = filter
-        self.filter_values = filter_values
-        self.action = action
-        self.action_field = action_field
-        self.action_value = action_value
